% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowPlot.R
\name{elbowPlot}
\alias{elbowPlot}
\title{elbowPlot}
\usage{
elbowPlot(estimation, title = "")
}
\arguments{
\item{estimation}{TODO}

\item{title}{the title of the plot}
}
\value{
A elbow plot
}
\description{
the Elbow plot permits to identify a subset of top-$k$ objects in combination with the heatmapPlot function. Objects are ordered according to their rank position given on the x-axis. On the y-axis we have the corresponding estimated signal value of each object. The idea of the elbow plot is to scan for 'jumps' - neighbouring signal estimates which are visually much distant - in an exploratory manner. The elbowPlot function requires the estimation results from the estimateTheta function.
}
\keyword{elbowtPlot}
