% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatedSignal.R
\name{estimatedSignal}
\alias{estimatedSignal}
\title{Object returned by the estimateTheta() function.}
\format{
A list of various values returned by the estimateTheta() function.
\describe{
 \item{estimation}{A data frame with the signal estimation and the standard error computed by the bootstrap for each object}
  \item{estimatedMatrixNoise}{The estimated matrix noise} 
  \item{allBootstraps}{The signal estimates from all bootstrap iterations}
}
}
\description{
Object returned by the estimateTheta() function.
}
\examples{
data(estimatedSignal)
}
