% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTheta.R
\name{estimateTheta}
\alias{estimateTheta}
\title{Estimation of the underlying signal.}
\usage{
estimateTheta(
  R.input,
  b,
  num.boot,
  solver,
  type,
  bootstrap.type,
  nCore = ((detectCores() - 1))
)
}
\arguments{
\item{R.input}{A matrix where the number of rows is equal to the number of the objects and
the number of columns equals the number of assessors (rankers).}

\item{b}{the penalization term. The suggested value is 0.1.}

\item{num.boot}{the number of boostrap samples created from the input ranked matrix. A positive number is expected.}

\item{solver}{A string that indicate which solver to use. Two options are available, 'gurobi' and 'nloptr'.
We recommend to use gurobi for faster computation. Note, a licence is required. Check the corresponding documentation on how to install gurobi.}

\item{type}{A string that indicates which model to use: four approaches are available: 'restrictedQuadratic', 'fullQuadratic', 'restrictedLinear' and 'fullLinear'.}

\item{bootstrap.type}{A string that indicate which bootstrap method to use: 'classic.bootstrap' or 'poisson.bootstrap'.}

\item{nCore}{the number of cores used for computation. Each core is used to solve a bootstrap sample. Default number is detectCores() - 1.}
}
\value{
A list with the estimation information obtained:
 \itemize{
  \item estimation - A data frame with the signal estimation and the standard error evaluating by the bootstrap for each object
  \item estimatedMatrixNoise - The estimated matrix noise 
  \item time - the execution time of the procedure
  \item allBootstraps - The single bootstrap estimations of all objects for each bootstrap sample
}
}
\description{
This is the main function for the estimation of the underlying latent signals. The parameters required are: (1) A rank matrix (objects as rows, assessors as columns), (2) the number of bootstrap samples (500 are recommended), and (3) a constant for the support variables \(b>0\), default is 0.1, (4)  the type of model with four different options: fullLinear, fullQuadratic, restrictedLinear, and restrictedQuadratic, (5) the type of  bootstrap sampling scheme, poisson.bootstrap and classic.bootstrap, and (6) the number of cores for parallel computation. Each bootstrap sample is executed on a dedicated CPU core.
}
\keyword{estimateTheta}
