% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapPlot.R
\name{heatmapPlot}
\alias{heatmapPlot}
\title{Heatmap noise matrix plot}
\usage{
heatmapPlot(estimation, type = "full", title = "")
}
\arguments{
\item{estimation}{The bootstrap estimation obtained from the estimateTheta function}

\item{type}{The type of method used: Two options are available, 'full' or 'reduced'}

\item{title}{The title of the plot}
}
\value{
A list with:
\itemize{
  \item plot - A heatmap plot with the noise matrix (ordered values).
  \item matrixNoiseOrdered - The matrix noise ordered by the columns. The objects are ordered by the estimated value.
  \item estimateThetaOrdered - The theta vector ordered by their importance (from the highest value to the lowest).
}
}
\description{
The heatmap plot allows us to check whether specific patterns are present in the data. The heatmapPlot contains the information about the involved noises in the estimation process. The rows of the noise matrix are ordered by the estimated rank of the signal value. The columns are ordered by the column sum. The column with the lowest sum is positioned on the left side and the column with the highest sum is positioned on the right side. Hence, the assessors positioned on the left should be more reliable than those on the right. When some assessors of low ranking ability are suspected the heatmap is very helpful. This plot is also useful when an informative set of top-ranked objects is likely or in general when patterns inside the data might be present. The heatmapPlot function requires the estimation results obtained from the estimateTheta function.
}
\keyword{heatmapPlot}
