% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlot}
\alias{violinPlot}
\title{violinPlot}
\usage{
violinPlot(estimation, trueSignal = NULL, title = NULL)
}
\arguments{
\item{estimation}{The estimation list from the 'estimateTheta' function}

\item{trueSignal}{The true signal (if available)}

\item{title}{The title of the plot}
}
\value{
A violint plot with the estimated distribution of each object
}
\description{
The violint plot displays the bootstrap distribution of the estimated signals along with the mean estimates. The deviations from the mean values are equal to 2+-SE and are displayed in the violin plot.  With the analysis of the distribution and the standard error of the signal of each object, it is possible to verify its rank stability compared to all other objects. The violinPlot function requires (1) the result obtained by the estimation procedure (see estimateTheta()) and (2) the 'true' (simulated) signals or ground truth (when available).
}
\keyword{violintPlot}
