% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elbowPlot.R
\name{elbowPlot}
\alias{elbowPlot}
\title{elbowPlot}
\usage{
elbowPlot(estimation, title = "")
}
\arguments{
\item{estimation}{Results from the estimateTheta() function}

\item{title}{A title for the plot}
}
\value{
A elbow plot
}
\description{
The elbow plot permits the identification of subsets of objects, e.g. top-$k$ or bottom-$q$ objects. On the x-axis all objects are ordered according to their rank positions. On the y-axis the corresponding estimated signal values are displayed. The idea of the elbow plot is to scan for 'jumps' in the sequence of ordered objects ? i.e. find signal estimates next to each other that are visually much distant - in an exploratory manner. The elbowPlot function requires the estimation results from the estimateTheta function.
}
\examples{
data(estimatedSignal)
elbowPlot(estimatedSignal)
}
\keyword{elbowtPlot}
