% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinPlot.R
\name{violinPlot}
\alias{violinPlot}
\title{violinPlot}
\usage{
violinPlot(estimation, trueSignal = NULL, title = NULL)
}
\arguments{
\item{estimation}{The estimation list from the 'estimateTheta' function}

\item{trueSignal}{The true signal (if available)}

\item{title}{The title of the plot}
}
\value{
A violint plot with the estimated distribution of each object
}
\description{
violinPlot
}
\examples{
data(estimatedSignal)
violinPlot(estimatedSignal)
}
\keyword{violintPlot}
